+++ CHANGELOG.md
@@ -33,6 +33,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - References to "Room Config" changed to "System Config"
  - Telemetry and Multi-Krang restructured into Services and ServiceProviders
 
+## [14.3.0] - 2020-10-06
+### Changed
+ - Loading/Starting Krang now runs StartSettings on the core
+ - KrangBootstrap Start has a callback for post-load but pre-start actions
+ - KrangControlSystem (Crestron) uses KrangBootstrap post-load callback to set program as started
+
 ## [14.2.1] - 2020-05-06
 ### Changed
  - Reverted "PrintVersions" console command back to "VersionInfo"